// src/types/partner.ts
export interface PartnerOnboardingWizardData {
  password: string;
  confirmPassword: string;
  phone?: string;
  company_name?: string;
  position_title?: string;
  organization_type?: string;
  organization_registration?: string;
  organization_website?: string;
  estimated_monthly_referrals?: string;
  payout_beneficiary_name?: string;
  payout_iban?: string;
  region?: string;
}